Refactor shared fixtures into conftest.py
